requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
distro==1.9.0
fastapi==0.124.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.9.0
pika==1.3.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1

# Celery + Redis + Gevent
celery==5.4.0
redis==5.2.0
kombu==5.4.2
gevent==24.11.1

# Tests
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Monitoring Docker (pour tests de charge)
docker==7.1.0